+++ packages/@tinacms/cli/src/next/config-manager.ts
@@ -11,7 +11,7 @@ import { loadProjectConfig } from '../next/vite'
 import { loadConfig } from 'tsconfig-paths'
 import { logger } from '../logger'
 import { resolveTsPathsToEsbuildAliases } from '../utils/alias-helpers'
-import { aliasPath } from '../utils/lib/esbuild-plugin-alias-path'
+import { aliasPath } from '../utils/esbuild-alias-lib/esbuild-plugin-alias-path'
 
 export const TINA_FOLDER = 'tina'
 export const LEGACY_TINA_FOLDER = '.tina'
