+++ packages/@tinacms/cli/src/utils/esbuild-alias-lib/normalize-options.ts
